[VOID] Timeline UI updates
Centered the Current Time/Frame display

The Frame display is now a QLabel with Green Highlight accents to match overall UI

Shifted timeline related code into timeline directory to keep things segregated as we grow

Add ToolTip formatter for Quick tooltip text generation with Header/Name: Description
