+++ src/VoidUi/QExtensions/Tooltip.h
@@ -1,0 +1,16 @@
+#ifndef _VOID_Q_EXT_TOOLTIP_H
+#define _VOID_Q_EXT_TOOLTIP_H
+
+/* STD */
+#include <string>
+
+/* Internal */
+#include "Definition.h"
+
+VOID_NAMESPACE_OPEN
+
+std::string ToolTipString(const std::string& name, const std::string& description);
+
+VOID_NAMESPACE_CLOSE
+
+#endif // _VOID_Q_EXT_TOOLTIP_H
